refactor: 단위테스트 각 클래스마다 하는 식으로 변경 작업
